+++ src/components/Style.css
@@ -1,5 +1,3 @@
-/** @format */
-
 .pink {
   background-color: #e74987;
 }
