--- src/testAPI.http
@@ -1,1 +1,0 @@
-GET https://pro-api.coinmarketcap.com/v1/cryptocurrency/listings/latest
